A bunch of Request and repository code
There is way too much code here for a single commit, but I don't really
care.

This adds a new Request class that wraps OkHttp, and creates a
repository class that can be created with different other classes to be
able to do the actual requests to Musicbrainz. They wrap the logic of
interacting with the API.