Shuffle around List interface
I have things that return the actual entity split from things that
return a EntityList (like RecordingList). I'm not sure if I love this,
but for now I'm going to keep it as it's a way to keep the concepts
separate. This might not be what the repositories actually return, but
it can be something that they can use.