Push encoding above Request
Based on some testing it seemed like having the encoding in the request
might have been causing some issues with the query. I'm not 100% sure,
but this is allowing it to work better for one of the queries that was
failing before.

I'm also replacing and stripping some of the values that were causing
issues where it was returning no results even though it should have.