Add artistname to query
Apparently for some artists you get better results when you use
artistname instead of just artist.

So I'm adding that as an option, and I'll likely use only that from now
on when querying.

Also adding a couple more other values that cause problems when querying
so I swap them out for nothing.