A whole bunch of Area entities
I had to make a specific AreaAdapter because of the relations. I haven't
done this with the other ones yet because I don't support their inner
relationships you can query, but for this one I had to because one of
the main reasons for querying an area is so you can find all related
areas, specifically the upper ones (a county for a city for example).