Allow adapters to be passed in
I want to allow for users of the ListSearch and IdSearch to provide
their own adapters. I haven't been able to get the Moshi code to work
with their JsonAdapter definitions on the classes, so I'm just putting
an Any here for now. I'll probably come revisit this in the future.